requirements.txt
upstox-python
python-dotenv
requests